% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{close.odbc32}
\alias{close.odbc32}
\title{Closes odbc32 connection}
\usage{
\method{close}{odbc32}(con)
}
\arguments{
\item{con}{odbc32 connection object}
}
\value{
logical
}
\description{
Closes odbc32 connection
}
\examples{
TBA
}
