% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{connect_remotely}
\alias{connect_remotely}
\title{Connect to Database Via ODBC using RODBC}
\usage{
connect_remotely(socket, FUN, ns = "RODBC", ...)
}
\arguments{
\item{socket}{socket}

\item{FUN}{function to be called remotely}

\item{\dots}{arguments to be passed to \code{FUN}}
}
\value{
odbc32 object
}
\description{
This helper function is used by \itemize{
  \item \code{\link{odbcConnectAccess}}
  \item \code{\link{odbcConnectAccess2007}}
  \item \code{\link{odbcConnect}}
  \item \code{\link{odbcDriverConnect}}
}
}
