% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc32.R
\name{create_odbc32_object}
\alias{create_odbc32_object}
\title{Create Object of Class "odbc32"}
\usage{
create_odbc32_object(socket, ref)
}
\arguments{
\item{socket}{socket}

\item{ref}{reference returned by \code{\link[r2r]{eval_remote}}}
}
\description{
Create Object of Class "odbc32"
}
