% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_tempfile.R
\name{data2access}
\alias{data2access}
\title{export data from an R object to MS Access db via separate 32-bit Rscript}
\usage{
data2access(data, accdb_file, ...)
}
\description{
export data from an R object to MS Access db via separate 32-bit Rscript
}
\details{
\describe{
  \item{overwrite}{logical; Attempt to drop any existing table of the name before writing data?}
  \item{convert_dates}{character; Convert dates as text before writing into database.}
}
}
