% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{dbFetch,Odbc32Result-method}
\alias{dbFetch,Odbc32Result-method}
\title{Retrieve records from Odbc32 query}
\usage{
\S4method{dbFetch}{Odbc32Result}(res, n = -1, ...)
}
\description{
Retrieve records from Odbc32 query
}
