% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{.append_con}
\alias{.append_con}
\title{Append a (new) connection to a global list of connections}
\usage{
.append_con(con)
}
\arguments{
\item{con}{}
}
\value{
a reference (index) number to the list of all connections
}
\description{
Append a (new) connection to a global list of connections
}
\examples{
#TBA
}
