% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc32.R
\docType{package}
\name{odbc32-package}
\alias{odbc32}
\alias{odbc32-package}
\title{odbc32: 32-bit ODBC drivers for 64-bit R sessions}
\description{
Tools for using 32-bit ODBC drivers from 64-bit R sessions. Based on socket connection between 64-bit R and host 32-bit R.
}
