% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{odbcConnect}
\alias{odbcConnect}
\title{establish ODBC connection defined by DSN name}
\usage{
odbcConnect(dsn, uid = "", pwd = "", socket = .GlobalEnv$.r2r_socket)
}
\arguments{
\item{dsn}{dsn}

\item{uid}{user name}

\item{pwd}{password}

\item{socket}{socket}
}
\description{
establish ODBC connection defined by DSN name
}
