% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{odbcConnectAccess}
\alias{odbcConnectAccess}
\title{Connect to an Access database (*.mdb)}
\usage{
odbcConnectAccess(access.file, uid = "", pwd = "",
  socket = .GlobalEnv$.r2r_socket)
}
\arguments{
\item{access.file}{path to .mdb file}

\item{uid}{user name}

\item{pwd}{password}
}
\value{
odbc32 object
}
\description{
Connect to an Access database (*.mdb)
}
\examples{
#TBA
}
