% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{odbcDataSources}
\alias{odbcDataSources}
\title{list data sources}
\usage{
odbcDataSources(type = c("all", "user", "system"), socket = default_socket())
}
\arguments{
\item{type}{see type param of RODBC::odbcDataSources()}
}
\description{
list data sources
}
