% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{odbcDriverConnect}
\alias{odbcDriverConnect}
\title{establish ODBC connection based on a connection string}
\usage{
odbcDriverConnect(..., socket = default_socket())
}
\arguments{
\item{...}{other args passed to RODBC::odbcDriverConnect()}

\item{socket}{}
}
\description{
establish ODBC connection based on a connection string
}
