% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{odbcEndTran}
\alias{odbcEndTran}
\title{clear/empty a db table}
\usage{
odbcEndTran(con, commit)
}
\arguments{
\item{con}{odbc32 connection object}

\item{commit}{logical}
}
\description{
clear/empty a db table
}
