% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_file}
\alias{package_file}
\title{Path to File in This Package}
\usage{
package_file(..., mustWork = FALSE)
}
\arguments{
\item{\dots}{passed to \code{\link{system.file}}}

\item{mustWork}{passed to \code{\link{system.file}}}
}
\value{
absolute file path to a file in the installation folder of this
  package
}
\description{
Path to File in This Package
}
