% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{print.odbc32}
\alias{print.odbc32}
\title{print method for odbc32}
\usage{
\method{print}{odbc32}(x)
}
\arguments{
\item{x}{odbc32 object}
}
\value{
printed odbc32 object
}
\description{
print method for odbc32
}
\examples{
#TBA
}
