% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_single.R
\name{remote_access_query}
\alias{remote_access_query}
\title{Query over remote connection}
\usage{
remote_access_query(...)
}
\arguments{
\item{...}{is passed to r2r::do.call_remote}
}
\value{
whatever the call on the remote R session returns
}
\description{
Query over remote connection
}
\examples{
#TBA
}
