% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rscript_path}
\alias{rscript_path}
\title{Path to Rscript.exe (on Windows)}
\usage{
rscript_path(arch = "i386")
}
\arguments{
\item{arch}{architecture: "i386" for 32 bit or "x64" for 64 bit}
}
\value{
absolute file path to "Rscript.exe"
}
\description{
Path to Rscript.exe (on Windows)
}
