% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{getSqlTypeInfo}
\alias{getSqlTypeInfo}
\alias{setSqlTypeInfo}
\title{Specify or retrieve a mapping of R types to DBMS datatypes.}
\usage{
getSqlTypeInfo(con, driver)

setSqlTypeInfo(con, driver, value)
}
\arguments{
\item{con}{odbc32 object}

\item{value}{passed to \code{RODBC::setSqlTypeInfo}}
}
\value{
For setSqlTypeInfo none.
For getSqlTypeInfo with an argument, a named list. Without an argument, a data frame.
}
\description{
Specify or retrieve a mapping of R types to DBMS datatypes.
}
