% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlClear}
\alias{sqlClear}
\title{clear/empty a db table}
\usage{
sqlClear(con, name)
}
\arguments{
\item{con}{odbc32 connection object}

\item{name}{name of db table to be dropped}
}
\description{
clear/empty a db table
}
