% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlDrop}
\alias{sqlDrop}
\title{drop a db table}
\usage{
sqlDrop(con, name)
}
\arguments{
\item{con}{odbc32 connection object}

\item{name}{name of db table to be dropped}
}
\description{
drop a db table
}
