% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlFetch}
\alias{sqlFetch}
\alias{sqlFetchMore}
\title{fetch db table}
\usage{
sqlFetch(con, name, ...)

sqlFetchMore(con, ...)
}
\arguments{
\item{con}{odbc32 connection object}

\item{name}{name of db table to be dropped}

\item{...}{other args passed to RODBC::sqlFetch() on remote}
}
\value{
data.frame
}
\description{
fetch db table
}
