% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlQuery}
\alias{sqlQuery}
\title{send query to a db and get results}
\usage{
sqlQuery(con, query, errors = NULL, ...)
}
\arguments{
\item{con}{odbc32 connection object}

\item{query}{text of the query}

\item{...}{other args passed to RODBC::sqlQuery() on remote}
}
\description{
send query to a db and get results
}
