% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlSave}
\alias{sqlSave}
\title{upload table to a db}
\usage{
sqlSave(con, data, name = NULL, ...)
}
\arguments{
\item{con}{odbc32 object}

\item{data}{tabular data}

\item{name}{name of the resulting db table}

\item{...}{other args passed to RODBC::sqlSave() on remote}
}
\description{
upload table to a db
}
