% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlTables}
\alias{sqlTables}
\title{list tables in DB}
\usage{
sqlTables(con, ...)
}
\arguments{
\item{con}{odbc32 connection object}

\item{...}{other args passed to RODBC::sqlTables() on remote}
}
\description{
list tables in DB
}
