% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{sqlUpdate}
\alias{sqlUpdate}
\title{update db table}
\usage{
sqlUpdate(con, data, name = NULL, ...)
}
\arguments{
\item{con}{odbc32 connection object}

\item{data}{tabular data object}

\item{name}{name of the resulting db table}

\item{...}{other args passed to RODBC::sqlUpdate() on remote}
}
\description{
update db table
}
