% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_single.R
\name{start_server_access2007}
\alias{start_server_access2007}
\title{Start r2r server connected to an Access database over 32-bit ODBC driver}
\usage{
start_server_access2007(accdb_fp, invisible = FALSE, wait = FALSE,
  address = "tcp://localhost", port = "5555", arch = "i386",
  Rbin = file.path(R.home(), "bin", arch, "Rscript.exe"))
}
\description{
Start r2r server connected to an Access database over 32-bit ODBC driver
}
