% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{stop_server}
\alias{stop_server}
\title{Stop r2r server running on a remote R session}
\usage{
stop_server(...)
}
\arguments{
\item{...}{args passed to r2r::stop_remote()}
}
\value{
logical
}
\description{
Stop r2r server running on a remote R session
}
\examples{
#TBA
}
