% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_tempfile.R
\name{write_access_db}
\alias{write_access_db}
\title{export data from Rdata file into MS Access db via separate 32-bit Rscript}
\usage{
write_access_db(rdata_file)
}
\description{
export data from Rdata file into MS Access db via separate 32-bit Rscript
}
