% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_multiple.R
\name{start_server}
\alias{start_server}
\title{Start a new R session with a running r2r server}
\usage{
start_server(address = "tcp://localhost", port = "5555",
  arch = "i386", Rbin = file.path(R.home(), "bin", arch,
  "Rscript.exe"), r_args = NULL, invisible = FALSE, wait = FALSE,
  global = TRUE, debug = FALSE)
}
\arguments{
\item{address}{host address}

\item{port}{port}

\item{arch}{architecture}

\item{Rbin}{path to Rscript binary}

\item{r_args}{further arguments to Rscript command}

\item{invisible}{logical; visibility of the new session window}
}
\value{
socket connection (invisible)
}
\description{
Start a new R session with a running r2r server
}
\examples{
#TBA
}
